clients/python/pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm[toml]>=7"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
root = "../../"
git_describe_command= 'git describe --tags --abbrev=0 --dirty --match "sunra_client_v*"'
tag_regex = "^sunra_client_v(?P<version>.*)$"

[project]
name = "sunra_client"
dynamic = ["version"]
description = "Python client for sunra.ai"
readme = "README.md"
authors = [{ name = "Features & Labels <support@sunra.ai>" }]
requires-python = ">=3.8"
dependencies = [
    "httpx>=0.21.0,<1",
    "httpx-sse>=0.4.0,<0.5",
]

[project.optional-dependencies]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "sphinx-autodoc-typehints",
]
test = [
    "pytest",
    "pytest-asyncio",
    "pillow",
]
dev = [
    "sunra_client[docs,test]",
]

[project.urls]
homepage = "https://sunra.ai"
repository = "https://github.com/sunra-ai/sunra-python"

[tool.pytest.ini_options]
testpaths = "tests/"
asyncio_mode = "auto"

[tool.setuptools]
packages = ["sunra_client"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"sunra_client" = ["py.typed"]
